Clean up the repo and remove app IDs (#119)
>⚠️ NOTE: This leaves some "in-flight" apps alone to avoid conflicts
with WIPs. Will gather them up soon and move them to
sync-todo/v2/client/.

This is a much-needed cleanup based on
https://github.com/mongodb-university/realm-template-apps/pull/118 . The
main template apps (the ones people refer to as "the template app") are
now in a `sync-todo` directory with the following versions:

- **v0:** old PBS app
- **v1:** current flex sync app
- **v2:** upcoming flex sync app with more features

All other apps have been relegated to the **other** directory.

Many of the template source directories now follow this structure:

- **backend:** The backend configuration.
- **client:** The client(s) source code.
- **generated:** Contains generated code. Don't edit this directly!
- **bluehawk.sh:** Generally bluehawks whatever's in client/ and puts it
in generated/. This keeps the client source clean when consumed by the
backend or viewed on an artifact repo.
- **realm-template.{json|xml|plist}:** Copied after bluehawking to the
client(s) in the generated directory. This scrubs the app ID used in
development.

The artifact repos will now pull from the generated/ directories to push
to the respective artifact repos.

Co-authored-by: Nick Larew <nick.larew@10gen.com>