Update React Native version (#177)
Closes #173 

We have some React Native and package dependency weirdness in the
template app. In this PR, I created a new React Native app, then
migrated the old template app's code to it. This is much easier than
trying to navigate upgrading the React Native version.

One big thing that didn't make the transition: icons. They were broken
in the old app anyway. It's certainly possible to fix that, but in the
interest of time, I wanted to get this updated app published. We can
look at adding icons back in later.