(DOCSP-34632): Flutter: Update dependencies and the README (#181)
The linked Jira ticket: https://jira.mongodb.org/browse/DOCSP-34632

Calls for me to investigate the issue reported in the forum: [Flexible
Sync Flutter App does not appear to sync
properly](https://www.mongodb.com/community/forums/t/flexible-sync-sample-flutter-app-does-not-appear-to-sync-properly/255202)

I was not able to replicate this Sync issue. I'm able to create new
documents from both the device and in Atlas, change values, and see the
updated values reflected on the device. I can't speak to whether this
might have been an issue with an older version of the SDK, as I updated
dependencies before I attempted to replicate this error.

I did encounter the issue reported in #168 . I'm checking with
engineering to see if it is safe for us to populate a default value for
this field, or if that could potentially clobber their process that
auto-generates the values when a user downloads the template client via
the UI or CLI. Meanwhile, I've updated the README to add a step for
populating this value, and also describing the behavior that a user can
expect when they don't provide values for the placeholders.