Fix missing data explorer link in Kotlin (#184)
My fix in #182 didn't fix the missing data explorer link in Kotlin. I
didn't realize that the bluehawk script was replacing the contents of
the client's `atlasConfig.xml` with a sanitized version, so the change I
made in the client file was being overwritten by the sanitized one
without that field.

This PR updates the sanitized version and the generated app's
`atlasConfig.xml`.