(DOCSP-36597): C++ Template App, MVC style (#197)
Add a C++ template app using FTXUI and the C++ SDK.

Ignore the C++ directory when running the Bluehawk copy script. Bluehawk
has a problem with using `::state:` in C++ code. For more details, refer
to this issue on the Bluehawk repo:
https://github.com/mongodb-university/Bluehawk/issues/145

I used Bluehawk copy for all other files and manually copied the
`database_manager.cpp` where the code triggering this issue is located.

---------

Co-authored-by: Chris Bush <chris.bush@10gen.com>