(DOCSP-38891) Refactor to support email/password auth (#202)
This PR refactors existing code to enable an email/password auth flow.
The new structure is set up to easily add more auth types in the future.
- Refactor to support email/password auth
- Move Edge Server connection logic to endpoint routes in Express server
  - Add new auth endpoints in Express server
- Add a log in page to React client that will show until one of the auth
endpoints responds with successful Edge Server connection
- Rename /node-server/ to /express-server/
- Update README to reflect new `edgectl` use

### Added after initial reviews:
**React Client**
- Persist user information in browser local storage
- Make the call for Edge Server Todo items repeat every 5 seconds to
refresh data. This is a less-than-ideal way to implement the the
refresh, but it works. Alternatives include adding a refresh button
instead or building out a proper stream. The latter is more work than we
have time for.

**Express server**
- Add proper REST error codes when API requests fail.

**Both**
- Add better error handling and messaging for API calls. Show
notifications when calls fail.
- Generally clean up the code to use consistent standards

---------

Co-authored-by: Dachary <dc@dacharycarey.com>