C++ Template App Follow-Ups (#206)
- Handle auth errors
- Adjust button and window constraints so the button labels are less
likely to disappear
- Use a delegate to have the `HomeController` change its own state,
instead of letting `DatabaseManager` mutate its state, per this PR
review comment:
https://github.com/mongodb-university/realm-template-apps/pull/197#discussion_r1543176629

Note for reviewer: after making this change, IntelliSense is saying the
DatabaseManager constructor is never used, and is warning me that the
`_delegate` I added to the DatabaseManager may be null. I'm not getting
those warnings in the other places that have delegates
(AuthManager/AppController, for example). So I think I may have missed
something here - BUT it is all working properly.